requirements.txt
streamlit
PyPDF2
furhat-realtime-api
google-genai
python-dotenv